Move all ecobenefits code to treemap app (fixes #805)
* species.py: move to treemap
* init.py, views.py and util.py: merge into a new ecobenefits.py
* urls.py, models.py: merge into their counterparts in treemap
* tests.py: move to tests/ecobenefits.py
* update import statements

Also: remove unused URL and view ("eco/benefit/tree/" and tree_benefits_view)
