+++ opentreemap/treemap/tests/__init__.py
@@ -444,6 +444,7 @@ from middleware import *      # NOQA
 from json_field import *      # NOQA
 from units import *           # NOQA
 from management import *      # NOQA
+from ecobenefits import *   # NOQA
 from ui.basic import *        # NOQA
 from ui.map import *          # NOQA
 from ui.plot_detail import *  # NOQA
